internal: modularize base-core and base types
