+++ generators/base/index.ts
@@ -19,3 +19,4 @@
 
 export { default } from './generator.js';
 export { default as command } from './command.js';
+export type { Config, Options, Features } from './types.js';
