+++ generators/info/generator.ts
@@ -22,6 +22,7 @@
 import chalk from 'chalk';
 
 import BaseCoreGenerator from '../base-core/index.js';
+import type { Config } from '../base/index.js';
 import JSONToJDLEntityConverter from '../../lib/jdl/converters/json-to-jdl-entity-converter.js';
 import JSONToJDLOptionConverter from '../../lib/jdl/converters/json-to-jdl-option-converter.js';
 import type { JHipsterGeneratorFeatures, JHipsterGeneratorOptions } from '../base/api.js';
@@ -29,7 +30,6 @@ import { JHIPSTER_CONFIG_DIR, YO_RC_FILE } from '../generator-constants.js';
 import { applicationsLookup } from '../workspaces/support/applications-lookup.js';
 import type { Entity } from '../../lib/types/base/entity.js';
 import { getEntitiesFromDir } from '../base-application/support/index.js';
-import type { Config } from '../base-core/types.js';
 import { replaceSensitiveConfig } from './support/utils.js';
 
 const isInfoCommand = commandName => commandName === 'info' || undefined;
